requirements.txt
mediapipe
pythermalcomfort
openai==1.55.3
requests
rich
smbus; sys_platform == 'linux'
python-dotenv
opencv-python
google-generativeai